requirements.txt
matplotlib==3.10.3
numpy==2.2.6
pandas==2.3.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.7.0
seaborn==0.13.2
selenium==4.33.0
torch==2.7.1+cu128
